app/testpmd: fix indirect flow action list ID size
Testpmd maps indirect action lists handlers into application IDs.
Testpmd API defines indirect IDs as 32 bits values.

The patch fixes indirect IDs size in  ACTION_INDIRECT_LIST_HANDLE
and ACTION_INDIRECT_LIST_CONF tokens

Fixes: 72a3dec7126f ("ethdev: add indirect flow list action")
Cc: stable@dpdk.org

Acked-by: Ferruh Yigit <ferruh.yigit@amd.com>

+++ app/test-pmd/cmdline_flow.c
@@ -7383,13 +7383,13 @@ static const struct token token_list[] = {
 		.name = "handle",
 		.help = "indirect list handle",
 		.next = NEXT(next_ial, NEXT_ENTRY(INDIRECT_LIST_ACTION_ID2PTR_HANDLE)),
-		.args = ARGS(ARGS_ENTRY_ARB(0, sizeof(uintptr_t))),
+		.args = ARGS(ARGS_ENTRY_ARB(0, sizeof(uint32_t))),
 	},
 	[ACTION_INDIRECT_LIST_CONF] = {
 		.name = "conf",
 		.help = "indirect list configuration",
 		.next = NEXT(next_ial, NEXT_ENTRY(INDIRECT_LIST_ACTION_ID2PTR_CONF)),
-		.args = ARGS(ARGS_ENTRY_ARB(0, sizeof(uintptr_t))),
+		.args = ARGS(ARGS_ENTRY_ARB(0, sizeof(uint32_t))),
 	},
 	[INDIRECT_LIST_ACTION_ID2PTR_HANDLE] = {
 		.type = "UNSIGNED",
